+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 % erl-log-formatter changelog
 
 # Next Version
+## Misc
+- Switch from jsx to [erl-json](https://github.com/galdor/erl-json).
 
 # 1.1.0
 ## Bugs